legacy/pyproject.toml
[project]
name = "plantbox"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "dataclasses>=0.8",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "langchain[openai]>=0.3.27",
    "numpy>=2.3.3",
    "openai>=1.106.1",
    "pyaudio>=0.2.14",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "setuptools>=80.9.0",
    "typing-extensions>=4.15.0",
    "websockets>=15.0.1",
    "streamlit>=1.39.0",
    "streamlit-webrtc>=0.63.4",
    "ultralytics>=8.3.199",
    "lap>=0.5.12",
    "tensorrt>=10.13.3.9",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "streamlit-chat>=0.1.1",
    "httpx[http2]>=0.28.1",
]
